Use through model on judgment.judges to preserve ordering
This way, the order in which the judges are added to the decision in the
admin area is preserved when presented to the user. We don't have to
explicitly add an ordering field -- the editor must just enter the
judges in the correct order.

We need to tweak some of the older migrations to work with theh existing
data, since nothing is changing under the hood. Only Django's view of
the database and how it works with it is changing.