document nature is now required
This ensures that all documents have a descriptive nature, even those
without a subtype. For example, all Legislation objects now default to
"Act". This means we can rely on the nature as a database-indexed
description of the type of every document.

This makes it possible to facet on the nature. Before, we had to facet
on a weird mixture of doc_type and nature.

In future, nature will be preferred over doc_type in search results. But
we need to backfill the nature field in ES for that to work.