saved-documnt buttons updated everywhere on the page
this means we can have multiple buttons in different places. they are
all identified with .save-document-button and are updated simultaneously
with htmx's hx-oob-swap

it also means we don't need the wrapper around the saved-document stuff
any more, because the two elements rendered on the htmx request are now
both swapped out with hx-oob-swap, so the target element and wrapper
don't matter.