make multiple save-doc buttons on one page possible
All the modals and buttons are now identified by document id.

The search view loads saved document buttons in bulk and injects them
into the page.

Streamline the saved document button view so that it doesn't redirect,
which is slow. Instead, the button view decides which template to use.

Combine the "not logged in" template into the "create" template.