avoid double-counting nested content when average embeddings
when we break an AKN document into chunks, we do it recursively. This
gives us good coverage and semantic search across provision boundaries
as well as accuracy when finding provision-level matches.

When we're calculating the average of all a document's chunks to find
the document-level embedding, we don't want to include the repeated
(nested) chunks because it gives them more weight than they should have.
Instead, use the average of the top-level provisions.